LPD-27050 No need of a helper class now there is one case
